Inacabado
Player causar dano e receber.
Armadilha funcionando.
Nova mecânica do player para atacar.
Inimigos ambos agora tomam dano/morrem e causam dano.

+++ Assets/Objects/Scripts/Bones.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Bones : MonoBehaviour
 {
     private GameManager gameManager;
-    public float BoneValue;
+    public int BoneValue;
 
     private void Start()
     {
